abstracted cross plot svg
